backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic==0.40.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
aiofiles==24.1.0
python-multipart==0.0.19
gitpython==3.1.43
PyYAML==6.0.2
httpx==0.28.1
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
